+++ packages/wallet/src/components/dialogs/CreatePaymentPointerDialog.tsx
@@ -3,7 +3,8 @@ import { Dialog, Transition } from '@headlessui/react'
 import { Fragment } from 'react'
 import { Select } from '@/ui/forms/Select'
 import type { DialogProps } from '@/lib/types/dialog'
-import { Form, useZodForm } from '@/ui/forms/Form'
+import { Form } from '@/ui/forms/Form'
+import { useZodForm } from '@/lib/hooks/useZodForm'
 import { z } from 'zod'
 import { Input } from '@/ui/forms/Input'
