+++ packages/wallet/src/pages/account/create.tsx
@@ -3,7 +3,8 @@ import { AppLayout } from '@/components/layouts/AppLayout'
 import { PageHeader } from '@/components/PageHeader'
 import { useDialog } from '@/lib/hooks/useDialog'
 import { Button } from '@/ui/Button'
-import { Form, useZodForm } from '@/ui/forms/Form'
+import { Form } from '@/ui/forms/Form'
+import { useZodForm } from '@/lib/hooks/useZodForm'
 import { Input } from '@/ui/forms/Input'
 import { Select } from '@/ui/forms/Select'
 import type {
