+++ packages/wallet/src/pages/transfer/request.tsx
@@ -2,7 +2,8 @@ import { AppLayout } from '@/components/layouts/AppLayout'
 import { Button } from '@/ui/Button'
 import { z } from 'zod'
 import Image from 'next/image'
-import { Form, useZodForm } from '@/ui/forms/Form'
+import { Form } from '@/ui/forms/Form'
+import { useZodForm } from '@/lib/hooks/useZodForm'
 import { Input } from '@/ui/forms/Input'
 import { Badge } from '@/ui/Badge'
 import { TransferHeader } from '@/components/TransferHeader'
